+++ addon/routes/list-form.js
@@ -133,6 +133,19 @@ FlexberryObjectlistviewHierarchicalRouteMixin, {
 
         this.sorting = userSettingsService.getCurrentSorting(componentName);
         this.perPage = userSettingsService.getCurrentPerPage(componentName);
+        if (this.perPage !== params.perPage) {
+          if (params.perPage !== 5) {
+            this.perPage = params.perPage;
+            userSettingsService.setCurrentPerPage(componentName, undefined, this.perPage);
+          } else {
+            if (this.sorting.length === 0) {
+              this.transitionTo(this.currentRouteName, { queryParams: { sort: null, perPage: this.perPage || 5 } }); // Show page without sort parameters
+            } else {
+              this.transitionTo(this.currentRouteName, { queryParams: { perPage: this.perPage || 5 } });  //Reload current page and records (model) list
+            }
+          }
+        }
+
         let queryParameters = {
           modelName: modelName,
           projectionName: projectionName,